add support for record fields
